+++ packages/cldr-core/__tests__/api/numbers/decimals.test.ts
@@ -373,35 +373,35 @@ test('decimal parts', () => {
   let p = api.formatDecimalToParts('12345.1234', opts);
 
   expect(p).toEqual([
-    { type: 'digits', value: '12' },
+    { type: 'integer', value: '12' },
     { type: 'group', value: ',' },
-    { type: 'digits', value: '345' },
+    { type: 'integer', value: '345' },
     { type: 'decimal', value: '.' },
-    { type: 'digits', value: '123' }
+    { type: 'fraction', value: '123' }
   ]);
 
   opts.style = 'percent';
   p = api.formatDecimalToParts('-1.234', opts);
   expect(p).toEqual([
-    { type: 'sign', value: '-' },
-    { type: 'digits', value: '123' },
+    { type: 'minus', value: '-' },
+    { type: 'integer', value: '123' },
     { type: 'percent', value: '%' }
   ]);
 
   p = api.formatDecimalToParts('9999.999');
   expect(p).toEqual([
-    { type: 'digits', value: '9' },
+    { type: 'integer', value: '9' },
     { type: 'group', value: ',' },
-    { type: 'digits', value: '999' },
+    { type: 'integer', value: '999' },
     { type: 'decimal', value: '.' },
-    { type: 'digits', value: '999' }
+    { type: 'fraction', value: '999' }
   ]);
 
   p = api.formatDecimalToParts('9999.9999');
   expect(p).toEqual([
-    { type: 'digits', value: '10' },
+    { type: 'integer', value: '10' },
     { type: 'group', value: ',' },
-    { type: 'digits', value: '000' }
+    { type: 'integer', value: '000' }
   ]);
 });
 
@@ -411,63 +411,63 @@ test('decimal scientific parts', () => {
 
   p = api.formatDecimalToParts('5', { style: 'scientific' });
   expect(p).toEqual([
-    { type: 'digits', value: '5' }
+    { type: 'integer', value: '5' }
   ]);
 
   p = api.formatDecimalToParts('5', { style: 'scientific', minimumSignificantDigits: 2 });
   expect(p).toEqual([
-    { type: 'digits', value: '5' },
+    { type: 'integer', value: '5' },
     { type: 'decimal', value: '.' },
-    { type: 'digits', value: '0' }
+    { type: 'fraction', value: '0' }
   ]);
 
   p = api.formatDecimalToParts('123.456', { style: 'scientific', minimumSignificantDigits: 3 });
   expect(p).toEqual([
-    {type: 'digits', value: '1'},
+    {type: 'integer', value: '1'},
     {type: 'decimal', value: '.'},
-    {type: 'digits', value: '23'},
+    {type: 'fraction', value: '23'},
     {type: 'exponent', value: 'E'},
-    {type: 'sign', value: '+'},
-    {type: 'digits', value: '2'}
+    {type: 'plus', value: '+'},
+    {type: 'integer', value: '2'}
   ]);
 
   p = api.formatDecimalToParts('21', { style: 'scientific', minimumSignificantDigits: 2 });
   expect(p).toEqual([
-    {type: 'digits', value: '2'},
+    {type: 'integer', value: '2'},
     {type: 'decimal', value: '.'},
-    {type: 'digits', value: '1'},
+    {type: 'fraction', value: '1'},
     {type: 'exponent', value: 'E'},
-    {type: 'sign', value: '+'},
-    {type: 'digits', value: '1'}
+    {type: 'plus', value: '+'},
+    {type: 'integer', value: '1'}
   ]);
 
   p = api.formatDecimalToParts('1578000', { style: 'scientific', minimumSignificantDigits: 2 });
   expect(p).toEqual([
-    {type: 'digits', value: '1'},
+    {type: 'integer', value: '1'},
     {type: 'decimal', value: '.'},
-    {type: 'digits', value: '6'},
+    {type: 'fraction', value: '6'},
     {type: 'exponent', value: 'E'},
-    {type: 'sign', value: '+'},
-    {type: 'digits', value: '6'}
+    {type: 'plus', value: '+'},
+    {type: 'integer', value: '6'}
   ]);
 
   p = api.formatDecimalToParts('-1.234567', { style: 'scientific', minimumSignificantDigits: 4 });
   expect(p).toEqual([
-    {type: 'sign', value: '-'},
-    {type: 'digits', value: '1'},
+    {type: 'minus', value: '-'},
+    {type: 'integer', value: '1'},
     {type: 'decimal', value: '.'},
-    {type: 'digits', value: '235'}
+    {type: 'fraction', value: '235'}
   ]);
 
   p = api.formatDecimalToParts('-0.00012345', { style: 'scientific', minimumSignificantDigits: 3 });
   expect(p).toEqual([
-    {type: 'sign', value: '-'},
-    {type: 'digits', value: '1'},
+    {type: 'minus', value: '-'},
+    {type: 'integer', value: '1'},
     {type: 'decimal', value: '.'},
-    {type: 'digits', value: '23'},
+    {type: 'fraction', value: '23'},
     {type: 'exponent', value: 'E'},
-    {type: 'sign', value: '-'},
-    {type: 'digits', value: '4'}
+    {type: 'minus', value: '-'},
+    {type: 'integer', value: '4'}
   ]);
 });